updated tries & added scoring to matching

+++ MapperatorTests/Matching/TrieDataMatcherTests.cs
@@ -38,7 +38,7 @@ public class TrieDataMatcherTests {
 
         foreach (var wordPosition in result) {
             //var rhythmString = dataRhythmStrings[wordPosition.Value];
-            //Console.WriteLine(string.Join('-', Enumerable.Range(0, searchLength).Select(o => rhythmString.Span[(int)wordPosition.CharPosition + o])));
+            //Console.WriteLine(string.Join('-', Enumerable.Range(0, searchLength).Select(o => rhythmString.Span[wordPosition.CharPosition + o])));
             //Console.WriteLine(string.Join('-', query.ToArray()));
             Assert.IsTrue(WordPositionInRange(wordPosition));
             Assert.IsTrue(WordPositionInRange(wordPosition, searchLength));
@@ -47,7 +47,7 @@ public class TrieDataMatcherTests {
     }
 
     private MapDataPoint GetMapDataPoint(WordPosition<int> wordPosition, int offset = 0) {
-        return mapDataPoints[wordPosition.Value][(int) wordPosition.CharPosition + offset];
+        return mapDataPoints[wordPosition.Value][wordPosition.CharPosition + offset];
     }
 
     private bool WordPositionInRange(WordPosition<int> wordPosition, int offset = 0) {
